@@ -133,17 +103,20 @@ setup(
             'manhole-cli = manhole.cli:main',
         ]
     },
-    keywords=[
-        'debugging', 'manhole', 'thread', 'socket', 'unix domain socket'
-    ],
-    python_requires='>=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*, !=3.5.*',
+    keywords=['debugging', 'manhole', 'thread', 'socket', 'unix domain socket'],
+    python_requires='>=3.8',
     install_requires=[
+        # eg: "aspectlib==1.1.1", "six>=1.7",
     ],
+    extras_require={
+        # eg:
+        #   "rst": ["docutils>=0.11"],
+        #   ":python_version=="2.6"": ["argparse"],
+    },
     cmdclass={
         'build': BuildWithPTH,
         'easy_install': EasyInstallWithPTH,
         'install_lib': InstallLibWithPTH,
         'develop': DevelopWithPTH,
-        'genpth': GeneratePTH,
     },
 )
