@@ -1,6 +1,3 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
 import sphinx_py3doc_enhanced_theme
 
 extensions = [
@@ -19,7 +16,7 @@ master_doc = 'index'
 project = 'manhole'
 year = '2012-2022'
 author = 'Ionel Cristian Mărieș'
-copyright = '{0}, {1}'.format(year, author)
+copyright = f'{year}, {author}'
 version = release = '1.8.0'
 
 pygments_style = 'trac'
@@ -28,10 +25,10 @@ extlinks = {
     'issue': ('https://github.com/ionelmc/python-manhole/issues/%s', '#'),
     'pr': ('https://github.com/ionelmc/python-manhole/pull/%s', 'PR #'),
 }
-html_theme = "sphinx_py3doc_enhanced_theme"
+html_theme = 'sphinx_py3doc_enhanced_theme'
 html_theme_path = [sphinx_py3doc_enhanced_theme.get_html_theme_path()]
 html_theme_options = {
-    'githuburl': 'https://github.com/ionelmc/python-manhole/'
+    'githuburl': 'https://github.com/ionelmc/python-manhole/',
 }
 
 html_use_smartypants = True
@@ -40,7 +37,7 @@ html_split_index = False
 html_sidebars = {
     '**': ['searchbox.html', 'globaltoc.html', 'sourcelink.html'],
 }
-html_short_title = '%s-%s' % (project, version)
+html_short_title = f'{project}-{version}'
 
 napoleon_use_ivar = True
 napoleon_use_rtype = False
